[Fix] Separate login / search data component
