+++ Raemin/CloneCoding/app/src/main/java/com/example/deliveryclonecoding/data/local/token/LocalTokenMapper.kt
@@ -1,6 +1,6 @@
 package com.example.deliveryclonecoding.data.local.token
 
-import com.example.deliveryclonecoding.data.remote.login.LoginDataItem
+import com.example.deliveryclonecoding.data.remote.login.datasource.LoginDataItem
 
 object LocalTokenMapper {
     fun mappingRemoteDataToLocal(loginDataItem: LoginDataItem): LocalTokenItem {
