--- Raemin/CloneCoding/app/src/main/java/com/example/deliveryclonecoding/data/remote/login/LoginDataSource.kt
@@ -1,7 +1,0 @@
-package com.example.deliveryclonecoding.data.remote.login
-
-import io.reactivex.Single
-
-interface LoginDataSource {
-    fun login(id: String, password: String): Single<LoginDataItem>
-}
