+++ Raemin/CloneCoding/app/src/main/java/com/example/deliveryclonecoding/data/remote/login/LoginRepositoryImpl.kt
@@ -1,15 +1,16 @@
-package com.example.deliveryclonecoding.data
+package com.example.deliveryclonecoding.data.remote.login
 
 import com.example.deliveryclonecoding.data.local.token.LocalTokenDataSource
 import com.example.deliveryclonecoding.data.local.token.LocalTokenMapper.mappingRemoteDataToLocal
-import com.example.deliveryclonecoding.data.remote.login.LoginDataSource
+import com.example.deliveryclonecoding.data.remote.login.datasource.LoginDataSource
+import com.example.deliveryclonecoding.data.remote.login.LoginRepository
 import io.reactivex.Completable
 import javax.inject.Inject
 
-class RepositoryImpl @Inject constructor(
+class LoginRepositoryImpl @Inject constructor(
     private val localTokenDataSource: LocalTokenDataSource,
     private val loginDataSource: LoginDataSource
-) : Repository {
+) : LoginRepository {
     override fun login(id: String, password: String): Completable {
         return loginDataSource
             .login(id, password)
