+++ Raemin/CloneCoding/app/src/main/java/com/example/deliveryclonecoding/data/remote/login/datasource/LoginDataItem.kt
@@ -1,4 +1,4 @@
-package com.example.deliveryclonecoding.data.remote.login
+package com.example.deliveryclonecoding.data.remote.login.datasource
 
 data class LoginDataItem(
     val access: String,
