+++ Raemin/CloneCoding/app/src/main/java/com/example/deliveryclonecoding/data/remote/login/datasource/LoginInfo.kt
@@ -1,4 +1,4 @@
-package com.example.deliveryclonecoding.data.remote.login
+package com.example.deliveryclonecoding.data.remote.login.datasource
 
 data class LoginInfo(
     val username: String,
