+++ Raemin/CloneCoding/app/src/main/java/com/example/deliveryclonecoding/di/RemoteDataModule.kt
@@ -1,7 +1,7 @@
 package com.example.deliveryclonecoding.di
 
-import com.example.deliveryclonecoding.data.remote.login.LoginDataSource
-import com.example.deliveryclonecoding.data.remote.login.LoginDataSourceImpl
+import com.example.deliveryclonecoding.data.remote.login.datasource.LoginDataSource
+import com.example.deliveryclonecoding.data.remote.login.datasource.LoginDataSourceImpl
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
