+++ Raemin/CloneCoding/app/src/main/java/com/example/deliveryclonecoding/di/RepositoryModule.kt
@@ -1,7 +1,7 @@
 package com.example.deliveryclonecoding.di
 
-import com.example.deliveryclonecoding.data.Repository
-import com.example.deliveryclonecoding.data.RepositoryImpl
+import com.example.deliveryclonecoding.data.remote.login.LoginRepository
+import com.example.deliveryclonecoding.data.remote.login.LoginRepositoryImpl
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -11,5 +11,5 @@ import dagger.hilt.components.SingletonComponent
 @Module
 abstract class RepositoryModule {
     @Binds
-    abstract fun bindsRepository(impl: RepositoryImpl): Repository
+    abstract fun bindsRepository(impl: LoginRepositoryImpl): LoginRepository
 }
